pyproject.toml
[project]
name = "model-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "numpy>=2.4.4",
    "onnxruntime>=1.24.4",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.2.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.24",
    "tensorflow>=2.21.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "ultralytics>=8.4.33",
]
